- Show fragment method added to login helper to be re-used
